backend/requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
jinja2>=3.1,<4.0
python-multipart>=0.0.20,<1.0
itsdangerous>=2.2,<3.0
google-generativeai>=0.8,<1.0
python-dotenv>=1.0,<2.0
